pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = [
    "/tests",
]

[project]
name = "xmessagebus"
version = "0.1"
authors = [
    { name="Yisu Peng", email="yisupeng@gmail.com" },
]
description = "A package to publish and receive messages"
#readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "absl-py",
    "xasyncio",
]

[project.urls]
"Homepage" = "https://github.com/shawn-peng/xmessagebus"
"Bug Tracker" = "https://github.com/shawn-peng/xmessagebus/issues"

#[tool.hatch.build.targets.wheel]
#packages = ["src/xmessagebus"]


